AddBusinessTransactionUseCase and supporting interfaces and classes
The first cut of this code which is likely to be updated in time since
for example update operations are not within the scope of this
usecase.

The persistence layer and REST layer should be done next before moving
onto new use cases. Likewise adding child scheduled transactions should
be added before supporting update use-cases.